Refactoring with Jetpack Compose - Part 4: MainActivity and Settings Page.

+++ app/src/main/java/com/crossbowffs/quotelock/data/api/QuoteStyle.kt
@@ -3,13 +3,16 @@ package com.crossbowffs.quotelock.data.api
 import android.graphics.Typeface
 
 data class QuoteStyle(
-    val quoteSize: Float,
-    val sourceSize: Float,
+    val quoteSize: Int,
+    val sourceSize: Int,
+    // Font properties
     val quoteTypeface: Typeface?,
     val quoteStyle: Int,
     val sourceTypeface: Typeface?,
     val sourceStyle: Int,
+    // Quote spacing
     val quoteSpacing: Int,
+    // Layout padding
     val paddingTop: Int,
     val paddingBottom: Int,
 )
